+++ NEWS.md
@@ -1,3 +1,9 @@
+# pkgnet (dev)
+
+## NEW FEATURES
+
+* `CreatePackageReport()` now outputs an object of new class `PackageReport` (instead of a list of reporters). This object will let you interactively manipulate the included reporter objects to customize the report, and regenerate the report on demand. You can also instantiate and interact with a `PackageReport` object directly without `CreatePackageReport()`.
+
 # pkgnet 0.4.0
 
 ## NEW FEATURES